Hotfix: Modularised query-timestamp test cases in its own file
